Implement simple client and fix bugs in server.
Besides a few bug fixes in the server, this commit adds the base
code for a very simple command line client. The client code must be
refactored before any additional development can be done here.

More code should be moved into common.py, so that is doesn't have to
be rewritten for every client.